Add save questions action and integrate with save subject
