+++ web/portal/src/constants/ActionTypes.js
@@ -24,6 +24,8 @@ export const COPY_PAGE = 'COPY_PAGE';
 export const EXCHANGE_PAGE = 'EXCHANGE_PAGE';
 export const EDIT_PAGE_TITLE = 'EDIT_PAGE_TITLE';
 export const DELETE_PAGE = 'DELETE_PAGE';
+export const RECIEVE_QUESTIONS_SUCCESS = 'RECIEVE_QUESTIONS_SUCCESS';
+export const RECIEVE_QUESTIONS_FAILURE = 'RECIEVE_QUESTIONS_FAILURE';
 
 // edit question
 export const SET_EDITQUESTION = 'SET_EDITQUESTION';
